\name{MicroNiche-package}
\alias{MicroNiche-package}
\alias{MicroNiche}
\docType{package}
\title{
\packageTitle{MicroNiche}
}
\description{
\packageDescription{MicroNiche}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MicroNiche}
\packageIndices{MicroNiche}

}
\author{
\packageAuthor{MicroNiche}

Maintainer: \packageMaintainer{MicroNiche}
}
\references{
Feinsinger et al. 1981. A simple measure of niche breadth. Ecology 62(1):27-32

Ludwig and Reynolds. 1988. Statistical Ecology. Wiley and Sons, Milton Australia
}

\keyword{ niche }
\seealso{
levins.Bn
hurlberts.Bn
feinsingers.PS
levins.overlap
proportional.overlap
}
\examples{
NA
}
